Fixed several chance bugs. The game is in English now.
Note this is a copy of Ciweilization 1.17 PC version instead of a working android game. This is just a starting point.

I originally wanted to use the last commit as the starting point of Ciweilization android development but that version contained frequently-occurring bugs and half of it were in Chinese, so I decided to fix those first.